pyproject.toml
[tool.poetry]
name = "teachingpy"
version = "0.1.0"
description = ""
authors = ["Iyed Ghedamsi <iyed.ghdemsi@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.7"
matplotlib = "^3.1.3"
palettable = "^3.3.0"
ipykernel = "^5.1.4"
numpy = "^1.18.1"
pulp = "^2.0"
sympy = "^1.5.1"
attrs = "^19.3.0"
seaborn = "^0.10.0"

[tool.poetry.dev-dependencies]
loguru = "^0.4.1"
pylama = "^7.7.1"
mypy = "^0.761"
ipdb = "^0.12.3"
black = {version = "^19.10b0", allow-prereleases = true}

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
